+++ dot.local/share/virtualenvs/globalworkspace-venv/bin/django-admin.py
@@ -1,0 +1,5 @@
+#!/home/mwinterschon/Projects/dotfiles-blumenkrieg/dot.local/share/virtualenvs/globalworkspace-venv/bin/python3.6m
+from django.core import management
+
+if __name__ == "__main__":
+    management.execute_from_command_line()